Check if conda already exists

+++ install_deps.sh
@@ -113,17 +113,19 @@ then
 	fi
 
 	if ! which conda 2> /dev/null; then
-		if y_or_n "Next: Install Conda"; then
-			wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
-			bash miniconda.sh -b -p $HOME/miniconda3
-			rm miniconda.sh
+		if [ ! -d "$HOME/miniconda3" ]; then
+			if y_or_n "Next: Install Conda"; then
+				wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
+				bash miniconda.sh -b -p $HOME/miniconda3
+				rm miniconda.sh
+			fi
 		fi
 	fi
 
 	# I won't ask the user first, because this is not a global installation.
 	# All of this stuff goes into a project-specific venv.
 	cd runner
-	$HOME/miniconda3/bin/conda env create -f environment.yml
+	$HOME/miniconda3/bin/conda env create --force -f environment.yml
 	cd ..
 else
 	echo "${0} does not support ${ID_LIKE} yet."
